Conflicts:
	extensions/GrandObjects/API/ThreadAPI.php

Former-commit-id: 3e5bc11c47bf93fd93d8c69af53b26312a267b41

+++ extensions/Reporting/Report/AbstractReport.php
@@ -1309,17 +1309,23 @@ abstract class AbstractReport extends SpecialPage {
             $src = $_FILES['image']['tmp_name'];
             $hash = md5($src);
             system("convert +antialias -background transparent $src /tmp/$hash.png");
-            list($width, $height) = getimagesize("/tmp/$hash.png");
-            $imgConst = DPI_CONSTANT*72/96;
-            $width = $width/$imgConst;
-            $height = $height/$imgConst;
-            $png = file_get_contents("/tmp/$hash.png");
-            unlink("/tmp/$hash.png");
-            $str = "top.$('input[aria-label=Width]').val($width);
-                    top.$('input[aria-label=Height]').val($height);
-                    top.$('.mce-btn.mce-open').parent().find('.mce-textbox').val('data:image/png;base64,".base64_encode($png)."').closest('.mce-window').find('.mce-primary').click();";
-            echo $str;
-            exit;
+            if(file_exists("/tmp/$hash.png")){
+                list($width, $height) = getimagesize("/tmp/$hash.png");
+                $imgConst = DPI_CONSTANT*72/96;
+                $width = $width/$imgConst;
+                $height = $height/$imgConst;
+                $png = file_get_contents("/tmp/$hash.png");
+                unlink("/tmp/$hash.png");
+                $str = "top.$('input[aria-label=Width]').val($width);
+                        top.$('input[aria-label=Height]').val($height);
+                        top.$('.mce-btn.mce-open').parent().find('.mce-textbox').val('data:image/png;base64,".base64_encode($png)."').closest('.mce-window').find('.mce-primary').click();";
+                echo $str;
+                exit;
+            }
+            else{
+                echo "alert('There was a problem with the uploaded file.  Make sure that it is a valid image and under 512KB.');";
+                exit;
+            }
         }
         return true;
     }
